src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/registration-context-delete-during-register-created-retrieval-expected.txt

Tests destroying a context during registration at the point when the created callback is retrieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS caught expected exception "TypeError: document.registerElement is not a function"
PASS child done
PASS successfullyParsed is true

TEST COMPLETE
